ai/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.6
h11==0.14.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
nltk==3.9.1
numpy==2.2.1
pandas==2.2.3
psutil==6.1.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
xgboost==2.1.3
